requirements.txt
py-clob-client>=0.34.6
python-dotenv>=1.0.0
eth-account>=0.13.0
eth-utils>=4.1.1
requests>=2.31.0
websockets>=12.0
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23
